requirements.txt
tensorflow==2.16.1+nv24.8
tensorrt==10.3.0
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
torch-tensorrt==2.8.0+cu126
torchao==0.14.0
torchcodec==0.6.0
numpy==1.26.4
cupy-cuda12x==13.6.0
flash-attn==2.8.4
xformers==0.0.33+ac00641.d20250830
triton==3.4.0
bitsandbytes==0.48.0.dev0
pycuda==2025.1.1
onnx==1.17.0
onnxruntime-gpu==1.23.0
protobuf==3.20.3
pytorch3d==0.7.8
kaolin==0.18.0
tinycudann==2.0
causal-conv1d==1.5.2
llama-cpp-python==0.3.16
ctranslate2==4.6.0
cuda-python==12.6.2.post1
nvidia-cudnn-frontend==1.14.0
nvidia-cutlass==4.0.0.0
pycolmap==3.13.0.dev0
pyceres==2.5
nerfacc==0.5.3
usd-core==25.11
cmake==3.31.6